+++ Reddit/Easy/making_links/src/lib.rs
@@ -1,0 +1,6 @@
+
+pub fn convert(input: &str) -> String {
+    println!("input: {}", input);
+
+    "".to_string()
+}