+++ Reddit/Easy/making_links/src/main.rs
@@ -1,0 +1,6 @@
+extern crate making_links;
+
+fn main() {
+    println!("{}", making_links::convert("7026425611433322325"));
+
+}